explicitly set project encoding
Since a while Eclipse wants to have us select an explicit encoding to
avoid depending on the default encoding of the operating system via the
workspace preferences (which is or was different on Windows and Linux).

Checking this in avoids a warning on recent IDE versions and ensures
that projects are using UTF-8 encoding both when being edited and when
compiled (where the maven encoding properties are relevant).