Move Maven configuration into pluginManagement section
It's generally a good practice to have as many configuration values as
possible in the pluginManagement section, as they will then surely be
used in all modules inheriting from this parent. If configuration values
are in the plugin section (or in a nested execution configuration), then
a second invocation of that same plugin in another Maven module might
not use the same configuration values.